+++ api-derive/parachains/info.ts
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.7';
-import type { CollatorId, ParaId } from 'https://deno.land/x/polkadot@0.2.13/types/interfaces/index.ts';
+import type { CollatorId, ParaId } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
 import type { DeriveApi, DeriveParachainActive, DeriveParachainFull, DeriveParachainInfo } from '../types.ts';
 import type { Active, DidUpdate, Heads, ParaInfoResult, PendingSwap, RelayDispatchQueue, RetryQueue, SelectedThreads } from './types.ts';
 
 import { map, of } from 'https://esm.sh/rxjs@7.5.7';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.13/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { memo } from '../util/index.ts';
 import { didUpdateToBool } from './util.ts';