+++ rpc-provider/bundle.ts
@@ -3,5 +3,5 @@
 
 export { HttpProvider } from './http/index.ts';
 export { packageInfo } from './packageInfo.ts';
-export { WsProvider } from './ws/index.ts';
 export { ScProvider } from './substrate-connect/index.ts';
+export { WsProvider } from './ws/index.ts';